Bugfix/runintransaction (#45)
* bugfix transaction issue

* rewrite runintransaction due some issues found in nested, switch, etc

---------

+++ kit_test.go
@@ -40,20 +40,24 @@ func TestRunInTransaction(t *testing.T) {
 		log.Fatal(err.Error())
 		t.Fail()
 	}
-	ctx = kit.ContextReset(ctx)
+	ctx, err = kit.ContextReset(ctx)
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
 	err = kit.RunInTransaction(
 		ctx,
-		func(ctx context.Context) error {
+		func(ctx context.Context) (context.Context, error) {
 			queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
 			_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('tiga')")
 			if err != nil {
-				return err
+				return ctx, err
 			}
 			_, err = queryable.ExecContext(ctx, "INSERT INTO `hello_2` VALUES('empat')")
 			if err != nil {
-				return err
+				return ctx, err
 			}
-			return nil
+			return ctx, nil
 		},
 	)
 	if err != nil {
@@ -61,6 +65,73 @@ func TestRunInTransaction(t *testing.T) {
 		t.Fail()
 	}
 }
+
+func TestConsecutiveRunInTransaction(t *testing.T) {
+	ctx := context.WithValue(context.Background(),
+		gosl.SQL_KEY,
+		gosl.NewQueryable(gosl.ConnectToDB(
+			"root",
+			"abcd",
+			"localhost",
+			"3306",
+			"test_1",
+			1,
+			1,
+			2*time.Minute,
+			2*time.Minute,
+		)))
+	kit := gosl.New(ctx)
+	queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
+	_, err := queryable.ExecContext(ctx, "DELETE FROM `hello`")
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
+	_, err = queryable.ExecContext(ctx, "DELETE FROM `hello_2`")
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
+	err = kit.RunInTransaction(
+		ctx,
+		func(ctx context.Context) (context.Context, error) {
+			queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
+			_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('tiga')")
+			if err != nil {
+				return ctx, err
+			}
+			_, err = queryable.ExecContext(ctx, "INSERT INTO `hello_2` VALUES('empat')")
+			if err != nil {
+				return ctx, err
+			}
+			return ctx, nil
+		},
+	)
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
+	err = kit.RunInTransaction(
+		ctx,
+		func(ctx context.Context) (context.Context, error) {
+			queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
+			_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('tiga')")
+			if err != nil {
+				return ctx, err
+			}
+			_, err = queryable.ExecContext(ctx, "INSERT INTO `hello_2` VALUES('empat')")
+			if err != nil {
+				return ctx, err
+			}
+			return ctx, nil
+		},
+	)
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
+}
+
 func TestRunInTransactionWithSwitchContext(t *testing.T) {
 	ctx := context.WithValue(context.Background(),
 		gosl.SQL_KEY,
@@ -95,33 +166,49 @@ func TestRunInTransactionWithSwitchContext(t *testing.T) {
 		log.Fatal(err.Error())
 		t.Fail()
 	}
-	queryable = kit.ContextSwitch(ctx, TKey).Value(gosl.SQL_KEY).(*gosl.Queryable)
-	_, err = queryable.ExecContext(ctx, "DELETE FROM `world`")
-	if err != nil {
-		log.Fatal(err.Error())
-		t.Fail()
-	}
-	ctx = kit.ContextReset(ctx)
-	err = kit.RunInTransaction(
-		ctx,
-		func(ctx context.Context) error {
-			queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
-			_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('sepuluh')")
+	if ctx, err := kit.ContextSwitch(ctx, TKey); err == nil {
+		if queryable, ok := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable); ok {
+			_, err = queryable.ExecContext(ctx, "DELETE FROM `world`")
 			if err != nil {
-				return err
+				log.Fatal(err.Error())
+				t.Fail()
 			}
-			queryable = kit.ContextSwitch(ctx, TKey).Value(gosl.SQL_KEY).(*gosl.Queryable)
-			_, err = queryable.ExecContext(ctx, "INSERT INTO `world` VALUES('empat')")
+			ctx, err = kit.ContextReset(ctx)
 			if err != nil {
-				return err
+				log.Fatal(err.Error())
+				t.Fail()
 			}
-			return nil
-		},
-	)
-	if err != nil {
-		log.Fatal(err.Error())
-		t.Fail()
+			err = kit.RunInTransaction(
+				ctx,
+				func(ctx context.Context) (context.Context, error) {
+					queryable, ok = ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
+					if !ok {
+						t.Fatal("failed to get queryable")
+					}
+					_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('sepuluh')")
+					if err != nil {
+						return ctx, err
+					}
+					if ctx, err = kit.ContextSwitch(ctx, TKey); err == nil {
+						queryable, ok = ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
+						if !ok {
+							t.Fatal("failed to get queryable")
+						}
+					}
+					_, err = queryable.ExecContext(ctx, "INSERT INTO `world` VALUES('empat')")
+					if err != nil {
+						return ctx, err
+					}
+					return ctx, nil
+				},
+			)
+			if err != nil {
+				log.Fatal(err.Error())
+				t.Fail()
+			}
+		}
 	}
+
 }
 
 func TestNestedRunInTransaction(t *testing.T) {
@@ -150,31 +237,34 @@ func TestNestedRunInTransaction(t *testing.T) {
 		log.Fatal(err.Error())
 		t.Fail()
 	}
-	ctx = kit.ContextReset(ctx)
+	ctx, err = kit.ContextReset(ctx)
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
 	err = kit.RunInTransaction(
 		ctx,
-		func(ctx context.Context) error {
+		func(ctx context.Context) (context.Context, error) {
 			queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
 			_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('tigabelas')")
 			if err != nil {
-				return err
+				return ctx, err
 			}
 			_, err = queryable.ExecContext(ctx, "INSERT INTO `hello_2` VALUES('empatbelas')")
 			if err != nil {
-				return err
+				return ctx, err
 			}
-
-			return kit.RunInTransaction(
+			err = kit.RunInTransaction(
 				ctx,
-				func(ctx context.Context) error {
+				func(ctx context.Context) (context.Context, error) {
 					queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
 					_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('tigabelasbelas')")
 					if err != nil {
-						return err
+						return ctx, err
 					}
-					return nil
-				},
-			)
+					return ctx, nil
+				})
+			return ctx, err
 		},
 	)
 	if err != nil {
@@ -209,31 +299,36 @@ func TestNestedRunInTransactionWithFailAtTheEnd(t *testing.T) {
 		log.Fatal(err.Error())
 		t.Fail()
 	}
-	ctx = kit.ContextReset(ctx)
+	ctx, err = kit.ContextReset(ctx)
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
 	err = kit.RunInTransaction(
 		ctx,
-		func(ctx context.Context) error {
+		func(ctx context.Context) (context.Context, error) {
 			queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
 			_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('satutigabelas')")
 			if err != nil {
-				return err
+				return ctx, err
 			}
 			_, err = queryable.ExecContext(ctx, "INSERT INTO `hello_2` VALUES('satuempatbelas')")
 			if err != nil {
-				return err
+				return ctx, err
 			}
 
-			return kit.RunInTransaction(
+			err = kit.RunInTransaction(
 				ctx,
-				func(ctx context.Context) error {
+				func(ctx context.Context) (context.Context, error) {
 					queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
 					_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('satutigabelasbelas')")
 					if err != nil {
-						return err
+						return ctx, err
 					}
-					return errors.New("fail deliberately")
+					return ctx, errors.New("fail deliberately")
 				},
 			)
+			return ctx, err
 		},
 	)
 	if err == nil {
